# web/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Machina API",
    "version": "1.0.0",
    "description": "REST API for Linux hypervisor hosts: QEMU/KVM domains via libvirt, networks, storage, host networking, backups, metrics, and optional KubeVirt bundle/apply workflows when configured."
  },
  "servers": [
    { "url": "/api/v1", "description": "Default API base" }
  ],
  "security": [
    { "bearerAuth": [] },
    { "cookieAuth": [] }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token authentication"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "machina_session",
        "description": "Cookie-based session authentication (`machina_session=`)"
      }
    },
    "schemas": {
...
      "KernelRouteChangeRequest": {
        "type": "object",
        "required": ["family", "operation", "destination"],
        "properties": {
          "family": { "type": "string", "enum": ["ipv4", "ipv6"] },
          "operation": { "type": "string", "enum": ["add", "delete"] },
          "destination": { "type": "string", "description": "default or CIDR / host prefix" },
          "via": { "type": "string", "nullable": true, "description": "Gateway IP for the route" },
          "dev": { "type": "string", "nullable": true, "description": "Output interface name" },
          "table": { "type": "integer", "format": "int32", "nullable": true, "description": "Routing table id (omit for main)" }
        }
      },
...
    }
  },
  "paths": {
...
    "/vms/console-info/{name}": {
      "get": {
        "summary": "Get console type info for a VM (VNC/SPICE/serial)",
        "tags": ["Console"],
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Console info" } }
      }
    },
...
    "/host/routing": {
      "post": {
        "summary": "Add or delete one validated kernel static route (ip route add|del; audited; browser session)",
        "tags": ["Host Networking"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/KernelRouteChangeRequest" } }
          }
        },
        "responses": {
          "200": { "description": "{ \"status\": \"ok\" }" },
          "400": { "description": "Invalid route parameters" },
          "403": { "description": "Browser session required" }
        }
      }
    },
...
    "/host/bridges": {
      "post": {
        "summary": "Create a network bridge",
        "tags": ["Host Networking"],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Bridge created" } }
      }
    },
    "/host/bridges/{name}": {
      "delete": {
        "summary": "Delete a network bridge",
        "tags": ["Host Networking"],
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Bridge deleted" } }
      }
    },
...
  }
}